start fixing exceptions
