+++ chakracore-cxx/lib/Common/Exceptions/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_library (Chakra.Common.Exceptions STATIC
-    # CommonExceptionsPch.cpp
     ExceptionCheck.cpp
     ExceptionBase.cpp
     ReportError.cpp
@@ -9,3 +8,8 @@ add_library (Chakra.Common.Exceptions STATIC
 # TODO (hanhossain): remove most directory paths
 # target_include_directories (
 #     Chakra.Common.Exceptions PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(Chakra.Common.Exceptions
+    PUBLIC include)
+
+target_link_libraries(Chakra.Common.Exceptions
+    PUBLIC Chakra.Common.Pal)
