+++ chakracore-cxx/lib/Common/Exceptions/ExceptionBase.cpp
@@ -2,7 +2,7 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
-#include "CommonExceptionsPch.h"
+#include "Exceptions/ExceptionBase.h"
 
 namespace Js
 {
