--- chakracore-cxx/lib/Common/Exceptions/OutOfMemoryException.h
@@ -1,13 +1,0 @@
-//-------------------------------------------------------------------------------------------------------
-// Copyright (C) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
-//-------------------------------------------------------------------------------------------------------
-#pragma once
-
-namespace Js {
-
-    class OutOfMemoryException : public ExceptionBase
-    {
-    };
-
-} // namespace Js
