+++ chakracore-cxx/lib/Common/Exceptions/ReportError.cpp
@@ -2,7 +2,6 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
-#include "CommonExceptionsPch.h"
 
 // Disable inline so that _ReturnAddress() will get the address of the calling function.
 _NOINLINE
