+++ chakracore-cxx/lib/Common/Exceptions/include/Exceptions/ExceptionBase.h
