+++ chakracore-cxx/lib/Common/Exceptions/include/Exceptions/OutOfMemoryException.h
@@ -4,9 +4,11 @@
 //-------------------------------------------------------------------------------------------------------
 #pragma once
 
+#include "Exceptions/ExceptionBase.h"
+
 namespace Js {
 
-    class NotImplementedException : public ExceptionBase
+    class OutOfMemoryException : public ExceptionBase
     {
     };
