+++ chakracore-cxx/lib/Common/Exceptions/include/Exceptions/RejitException.h
@@ -4,6 +4,8 @@
 //-------------------------------------------------------------------------------------------------------
 #pragma once
 
+#include "Exceptions/ExceptionBase.h"
+
 namespace Js
 {
     class RejitException : public ExceptionBase
