+++ chakracore-cxx/lib/Common/Exceptions/include/Exceptions/ScriptAbortException.h
@@ -2,5 +2,14 @@
 // Copyright (C) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE.txt file in the project root for full license information.
 //-------------------------------------------------------------------------------------------------------
-#include "CommonExceptionsPch.h"
+#pragma once
 
+#include "Exceptions/ExceptionBase.h"
+
+namespace Js {
+
+    class ScriptAbortException : public ExceptionBase
+    {
+    };
+
+} // namespace Js
